% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes}
\alias{aes}
\title{Title
mapping function like ggplot aes()}
\usage{
aes(g, x, y, color = NULL, size = NULL, shape = NULL, group = NULL)
}
\arguments{
\item{group}{}
}
\value{

}
\description{
Title
mapping function like ggplot aes()
}
