% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotations}
\alias{annotations}
\title{Annotation settings for g2plot}
\usage{
annotations(
  gWidget,
  type = c("text", "line", "image", "region", "dataMarker", "dataRegion",
    "regionFilter", "shape", "html"),
  position = c(NA, NA),
  animate = FALSE,
  autoAdjust = TRUE,
  style = NULL,
  start = NULL,
  end = NULL,
  src = NULL,
  content = NULL,
  rotate = NULL,
  maxLength = NULL,
  autoEllipsis = TRUE,
  lineLength = NULL,
  container = NULL,
  html = NULL,
  offsetX = NULL,
  offsetY = NULL,
  alignX = NULL,
  alignY = NULL,
  color = NULL
)
}
\arguments{
\item{type}{图形标注类型}

\item{position}{标注位置.}

\item{animate}{是否进行动画.}

\item{autoAdjust}{文本超出绘制区域时，是否自动调节文本方向.}

\item{src}{图片路径}

\item{content}{文本内容}
}
\description{
Annotation settings for g2plot
}
