% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{area}
\usage{
area(
  data,
  xField = NULL,
  yField = NULL,
  title = "",
  description = "",
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  padding = NULL,
  color = NULL,
  tooltip = NULL,
  legend = NULL,
  label = NULL,
  areaStyle = NULL,
  line = NULL,
  point = NULL,
  xAxis = NULL,
  yAxis = NULL,
  theme = NULL,
  events = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{required
数据源为对象集合，例如：[{ segment: '分类一'，value: 20 }, { segment: '分类二'，value: 20 }]。}

\item{xField}{required, string 类型
面积形状在 x 方向（横向延伸）对应的数据字段名，一般对应一个连续字段。}

\item{yField}{required, string 类型
面积形状在 y 方向对应的数据字段名，一般对应一个离散字段。}

\item{title}{optional 见通用图表配置。}

\item{description}{optional 见通用图表配置。}

\item{width}{optional 见通用图表配置。}

\item{height}{optional 见通用图表配置。}

\item{forceFit}{optional 见通用图表配置。}

\item{padding}{optional 见通用图表配置。}

\item{color}{optional, string[] | function 类型
指定折线颜色。如不进行配置则采用 theme 中的配色。}

\item{tooltip}{optional 见通用图表配置。}

\item{legend}{optional 见通用图表配置。}

\item{label}{optional
visible: boolean  图形标签是否显示
formatter: function  对 label 的显示文本进行格式化
offsetX: number 在 label 位置的基础上再往 x 方向的偏移量
offsetY: number 在 label 位置的基础上再往 y 方向的偏移量
style: object 配置 label 文本}

\item{areaStyle}{optional, object | function 类型
配置面积样式。
strokeStyle: string  面积的边框颜色
opacity: number  透明度
lineWidth: number  面积区域边框线的线宽
lineDash: number[]  面积区域边框线的虚线配置
另外还支持回调函数的配置方式，入参为当前图形的对应数据，出参为一个样式配置对象。}

\item{line}{optional, object 类型
配置面积图上线，起到辅助阅读的作用。分组及颜色映射方式与面积图形保持一致。
visible: boolean  是否显示数据线
opacity: number  透明度
color: string  颜色
size: number  线宽
style: object | function  线图形样式，另外还支持回调函数的配置方式，入参为当前图形的对应数据，出参为一个样式配置对象
point
optional, object 类型
配置面积图上的数据点，起到辅助阅读的作用。分组及颜色映射方式与面积图形保持一致。
visible: boolean  是否显示数据点
shape: string  数据点形状
size: number  数据点大小
style: object | function  数据点图形样式，另外该属性还支持回调函数的配置方式，入参为当前图形的对应数据，出参为一个样式配置对象}

\item{xAxis}{optional 见通用图表配置。}

\item{yAxis}{optional 见通用图表配置。}

\item{theme}{optional 见通用图表配置。}

\item{events}{optional
图形事件 onAreaClick: function 区域点击事件
onAreaDblclick: function 区域双击事件
onAreaMousemove: function 区域鼠标移动事件
onAreaMouseenter: function 区域鼠标进入事件
onAreaMouseleave: function 区域鼠标移出事件
onAreaMousedown: function 区域鼠标点击事件
onAreaMouseup: function 区域鼠标抬起事件
onAreaContextmenu: function 右键事件
如配置了区域图上的数据点：
onPointClick: function  数据点点击事件
onPointDblClick: function 数据点双击事件
onPointMousemove: function 数据点鼠标移动事件
onPointMouseenter: function 数据点鼠标进入事件
onPointMouseleave: function 数据点鼠标移出事件
onPointMousedown: function 数据点鼠标点击事件
onPointMouseup: function 数据点鼠标抬起事件
onPointContextmenu: function 数据点右键事件
如配置了区域图上的折线：
onLineClick: function 折线点击事件
onLineDblClick: function 折线双击事件
onLineMousemove: function 折线鼠标移动事件
onLineMouseenter: function 折线鼠标进入事件
onLineMouseleave: function 折线鼠标移出事件
onLineMousedown: function 折线鼠标点击事件
onLineMouseup: function 折线鼠标抬起事件
onLineContextmenu: function 折线右键事件
其他事件类型见通用图表配置。}
}
\description{
area
}
