% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Axis}
\alias{Axis}
\alias{xAxis}
\alias{yAxis}
\title{Create an axis for a chart}
\usage{
Axis(
  g2Htmlwidget,
  which = c("x", "y"),
  visible = TRUE,
  type = c("dateTime", "time", "linear", "time", "cat", "dateTime", "category", "log",
    "pow", "time"),
  tickCount,
  tickInterval,
  min,
  max,
  line = list(visible = TRUE, style = lineStlye()),
  grid = list(visible = TRUE, style = lineStlye()),
  label = list(visible = TRUE, suffix = NULL, precision = 2, offsetX = 0, offsetY = 0,
    autoHide = TRUE, autoRotate = TRUE, style = textStyle()),
  tickLine = list(visible = TRUE, style = lineStlye()),
  title = list(text = "", offset = 12, style = textStyle())
)

xAxis(g2Htmlwidget, ...)

yAxis(g2Htmlwidget, ...)
}
\description{
Add an axis to a chart.
}
