% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{<Add Title>
g2 function
<Add Description>
g function ,start a plot}
\usage{
g(
  data = NULL,
  plot_type = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  elementId = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{input data}

\item{ylab}{}
}
\description{
<Add Title>
g2 function
<Add Description>
g function ,start a plot
}
