% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{g2label}
\alias{g2label}
\title{Title}
\usage{
g2label(
  g2plot,
  type,
  offset = NULL,
  offsetX = NULL,
  offsetY = NULL,
  content = NULL,
  style = NULL,
  autoRotate = NULL,
  rotate = NULL,
  labelLine = NULL,
  labelEmit = NULL,
  layout = NULL,
  position = NULL,
  animate = NULL,
  formatter = NULL,
  autoHide = NULL
)
}
\arguments{
\item{autoHide}{}
}
\value{

}
\description{
Title
}
