% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2matrix.R
\name{g2matrix}
\alias{g2matrix}
\title{g2martrix 
g2martrix heatmap}
\usage{
g2matrix(
  data,
  xField = NULL,
  yField = NULL,
  colorField = NULL,
  color = NULL,
  title = "",
  description = "",
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  padding = NULL,
  sizeField = NULL,
  shapeSize = NULL,
  shapeType = NULL,
  forceSquare = NULL,
  xAxis = NULL,
  yAxis = NULL,
  tooltip = NULL,
  legend = NULL,
  label = NULL,
  theme = NULL,
  background = NULL,
  events = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{xField}{}

\item{yField}{}

\item{colorField}{}

\item{color}{}

\item{title}{}

\item{height}{}

\item{forceFit}{}

\item{padding}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{theme}{}

\item{events}{}
}
\description{
g2martrix 
g2martrix heatmap
}
