% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g2plot-shiny}
\alias{g2plot-shiny}
\alias{gOutput}
\alias{renderG}
\title{Shiny bindings for g2plot}
\usage{
gOutput(outputId, width = "100\%", height = "400px")

renderG(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a g2plot}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using g2plot within Shiny
applications and interactive Rmd documents.
}
