% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{heatmap 
plot heatmap}
\usage{
heatmap(
  data,
  xField = NULL,
  yField = NULL,
  colorField = NULL,
  color = NULL,
  title = "",
  description = "",
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  padding = NULL,
  point = FALSE,
  xAxis = NULL,
  yAxis = NULL,
  tooltip = NULL,
  legend = NULL,
  label = NULL,
  theme = NULL,
  background = NULL,
  events = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{xField}{}

\item{yField}{}

\item{colorField}{}

\item{color}{}

\item{title}{}

\item{height}{}

\item{forceFit}{}

\item{padding}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{theme}{}

\item{events}{}
}
\description{
heatmap 
plot heatmap
}
