% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Title label settings for g2plot}
\usage{
label(
  gWidget,
  type,
  offset = NULL,
  offsetX = NULL,
  offsetY = NULL,
  content = NULL,
  style = NULL,
  autoRotate = NULL,
  rotate = NULL,
  labelLine = NULL,
  labelEmit = NULL,
  layout = NULL,
  position = NULL,
  animate = NULL,
  formatter = NULL,
  autoHide = NULL
)
}
\arguments{
\item{type}{当用户使用了自定义的 label 类型，需要声明具体的 type 类型，否则会使用默认的 label 类型渲染（饼图 label 支持 inner|outer|spider）}

\item{autoHide}{}

\item{g}{object for g2plot}
}
\value{

}
\description{
Title label settings for g2plot
}
