% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend}
\alias{legend}
\title{Title}
\usage{
legend(
  g,
  legend = TRUE,
  layout = NULL,
  position = NULL,
  title = NULL,
  offsetX = NULL,
  offsetY = NULL,
  background = NULL,
  flipPage = NULL,
  itemWidth = NULL,
  itemHeight = NULL,
  itemName = NULL,
  itemValue = NULL,
  itemSpacing = NULL,
  animate = NULL,
  animateOptio = NULL,
  label = NULL,
  marker = NULL,
  maxWidth = NULL,
  maxHeight = NULL,
  rail = NULL,
  reversed = NULL,
  min = NULL,
  max = NULL,
  value = NULL,
  slidable = NULL,
  track = NULL,
  handler = NULL,
  custom = NULL,
  items = NULL
)
}
\arguments{
\item{items}{}
}
\value{

}
\description{
Title
}
