% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{line}
\alias{line}
\title{line}
\usage{
line(
  data,
  meta = NULL,
  xField = NULL,
  yField = NULL,
  seriesField = NULL,
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  pixelRatio = NULL,
  lineSize = NULL,
  lineStyle = NULL,
  smooth = NULL,
  color = NULL,
  point = FALSE,
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{xField}{}

\item{yField}{}

\item{seriesField}{}

\item{height}{}

\item{forceFit}{}

\item{lineSize}{}

\item{lineStyle}{}

\item{smooth}{}

\item{color}{}

\item{point}{}

\item{title}{}

\item{padding}{}

\item{theme}{}

\item{xAxis}{}

\item{yAxis}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{events}{}
}
\description{
line
}
