% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie}
\alias{pie}
\title{pie}
\usage{
pie(
  data,
  angleField = NULL,
  colorField = NULL,
  color = NULL,
  title = "",
  description = "",
  width = NULL,
  height = NULL,
  radius = NULL,
  forceFit = TRUE,
  padding = NULL,
  tooltip = NULL,
  legend = NULL,
  label = NULL,
  theme = NULL,
  events = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{colorField}{}

\item{color}{}

\item{title}{}

\item{height}{}

\item{forceFit}{}

\item{padding}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{theme}{}

\item{events}{}

\item{data:}{collection}

\item{radius:}{number}

\item{angleField:}{string}

\item{colorField:}{string}

\item{color:}{string | string[] | function}

\item{xField}{}

\item{yField}{}

\item{pointStyle}{}

\item{quadrant}{}

\item{trendline}{}
}
\description{
pie
}
