% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{rose-shiny}
\alias{rose-shiny}
\alias{roseOutput}
\alias{renderRose}
\title{Shiny bindings for rose}
\usage{
roseOutput(outputId, width = "100\%", height = "400px")

renderRose(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
strose and have \code{'px'} appended.}

\item{expr}{An expression that generates a rose}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using rose within Shiny
applications and interactive Rmd documents.
}
