% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{rose}
\alias{rose}
\title{rose}
\usage{
rose(
  data,
  meta = NULL,
  radiusField = NULL,
  catagoryField = NULL,
  colorField = NULL,
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  pixelRatio = NULL,
  redius = NULL,
  color = NULL,
  sectorStyle = c(fill = NULL, stroke = NULL, lineWidth = NULL, lineDash = NULL,
    opacity = NULL, fillOpacity = NULL, strokeOpacity = NULL),
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{height}{}

\item{forceFit}{}

\item{color}{}

\item{title}{}

\item{padding}{}

\item{theme}{}

\item{xField}{}

\item{yField}{}

\item{seriesField}{}

\item{roseSize}{}

\item{smooth}{}

\item{roseStyle}{}

\item{point}{}

\item{xAxis}{}

\item{yAxis}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{events}{}
}
\description{
rose
}
