% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(g, xField = NULL, yField = NULL, colorField = NULL)
}
\arguments{
\item{xField}{}

\item{yField}{}

\item{colorField}{}

\item{title}{}

\item{height}{}

\item{forceFit}{}

\item{padding}{}

\item{theme}{}

\item{data}{}

\item{color}{}

\item{pointStyle}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{quadrant}{}

\item{trendline}{}

\item{events}{}
}
\description{
scatter
}
