% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{shapestyle}
\alias{shapestyle}
\title{shape style settings}
\usage{
shapestyle(
  fill = NULL,
  r = NULL,
  fillOpacity = NULL,
  stroke = NULL,
  lineWidth = NULL,
  lineDash = NULL,
  lineOpacity = NULL,
  opacity = NULL,
  shadowColor = NULL,
  strokeOpacity = NULL,
  shadowBlur = NULL,
  shadowOffsetX = NULL,
  shadowOffSetY = NULL,
  cursor = NULL
)
}
\arguments{
\item{cursor}{}
}
\value{

}
\description{
shape style settings
}
