% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackbar.R
\name{stackbar}
\alias{stackbar}
\title{stackbar}
\usage{
stackbar(
  data,
  xField = NULL,
  yField = NULL,
  stackField = NULL,
  colorField = NULL,
  color = NULL,
  barSize = NULL,
  barStyle = NULL,
  title = "",
  description = "",
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  padding = NULL,
  xAxis = NULL,
  yAxis = NULL,
  tooltip = NULL,
  legend = NULL,
  label = NULL,
  theme = NULL,
  conversionTag = NULL,
  events = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{required
数据源为对象集合，例如：[{ segment: 分类一，value: 20 }, { segment: 分类二，value: 20 }]。}

\item{xField}{required, string 类型
条形在 x 方向长度映射对应的数据字段名，一般对应一个离散字段。}

\item{yField}{required, string 类型
条形在 y 方向位置映射所对应的数据字段名，一般对应一个分类字段。}

\item{colorField}{optional, string 类型
条形颜色映射对应的数据字段名。}

\item{color}{optional, string | string[] | function 类型
指定条形颜色。如不进行配置则采用 theme 中的配色。}

\item{barStyle}{optional, object | function 类型
配置条形样式。
fill: string  填充颜色
stroke: string  描边颜色
lineWidth: number  描边宽度
lineDash: number[]  描边虚线显示
opacity: number  透明度
另外还支持回调函数的配置方式，入参为当前图形的对应数据，出参为一个样式配置对象。}

\item{title}{optional 见通用图表配置。}

\item{description}{optional 见通用图表配置。}

\item{width}{optional 见通用图表配置。}

\item{height}{optional 见通用图表配置。}

\item{forceFit}{optional 见通用图表配置。}

\item{padding}{optional 见通用图表配置。}

\item{xAxis}{optional 见通用图表配置。}

\item{yAxis}{optional 见通用图表配置。}

\item{tooltip}{optional 见通用图表配置。}

\item{legend}{optional 见通用图表配置。}

\item{label}{optional
visible: boolean    图形标签是否显示
position: 'top' | 'middle' | 'bottom'    图形标签相对于柱形的位置
formatter: function  对 label 的显示文本进行格式化。
offsetX: number 在 label 位置的基础上再往 x 方向的偏移量。
offsetY: number 在 label 位置的基础上再往 y 方向的偏移量。
style: object 配置 label 文本样式}

\item{theme}{optional 见通用图表配置。}

\item{conversionTag}{optional
visible: boolean 转化率组件是否显示，默认 false。
size: number 转化率组件宽度，默认 80。
spacing: number 与条图形的间距，默认 12。
offset: number 相对 x 轴的偏移距离，默认 0。
arrow.visible: boolean 箭头图形是否显示，默认 true。
arrow.headSize: number 箭头图形尖的高度，默认 12。
arrow.style: object 箭头图形绘图属性。
value.visible: boolean 转化率值是否显示，默认 true。
value.style: object 转化率值绘图属性。
value.formatter: (v1, v2) => string 转化率值 formatter，用于自定义内容，v1 为上侧条的数值，v2 位下侧条的数值。}

\item{events}{optional
图形事件 onstackbarClick: function  条形点击事件
onstackbarDblClick: function  条形双击事件
onstackbarMousemove: function  条形鼠标移动事件
onstackbarContextmenu: function    条形右键事件
其他事件类型见通用图表配置。}

\item{barSize:}{number
optional, number 类型
设置条形的高度为一个固定值。}
}
\description{
stackbar
}
