% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{stock}
\alias{stock}
\title{Stock plot}
\usage{
stock(
  gWidget,
  width = NULL,
  height = NULL,
  open = "open",
  close = "close",
  high = "high",
  low = "low"
)
}
\description{
Stock plot
}
\examples{
g(data) \%>\% stock() \%>\% g2plot::aes(x ="trade_date",y = "y")
}
