% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{title}
\alias{title}
\title{Title setting for axis}
\usage{
title(
  text = NULL,
  offset = NULL,
  spacing = NULL,
  style = NULL,
  autoRotate = NULL
)
}
\arguments{
\item{text}{text content for xAxis}

\item{autoRotate}{}
}
\value{

}
\description{
Title setting for axis
}
