% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{Tooltip for plot
Tooltip, which means that when the mouse hovers over the chart or the finger clicks on a data point of the mobile device,
the data of the point is displayed in the form of interactive prompt information, such as the value of the point,
the data unit, etc.}
\usage{
tooltip(
  gWidget,
  title = NULL,
  fields = c(NULL, NULL),
  formatter = NULL,
  follow = TRUE,
  enterable = FALSE,
  showTitle = NULL,
  position = c("top", "bottom", "left", "right"),
  shared = NULL,
  showCrosshairs = FALSE,
  crosshairs = NULL,
  textStyle = NULL,
  textBackgroundStyle = NULL,
  showMarkers = TRUE,
  marker = NULL,
  showContent = NULL,
  container = NULL,
  containerTp,
  itemTpl,
  domStyles,
  offset,
  customContent
)
}
\arguments{
\item{customContent}{}
}
\value{

}
\description{
Tooltip for plot
Tooltip, which means that when the mouse hovers over the chart or the finger clicks on a data point of the mobile device,
the data of the point is displayed in the form of interactive prompt information, such as the value of the point,
the data unit, etc.
}
\examples{

}
