% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{xAxis}
\alias{xAxis}
\title{xAxis settings}
\usage{
xAxis(
  gWidget,
  position = NULL,
  label = NULL,
  min = NULL,
  max = NULL,
  minLimit = NULL,
  maxLimit = NULL,
  tickCount = NULL,
  tickInterval = NULL,
  tickMethod = NULL,
  line = NULL,
  tickLine = NULL,
  subTickLine = NULL,
  title = NULL,
  grid = NULL,
  animate = NULL
)
}
\arguments{
\item{gWidget}{}

\item{tickMethod}{tickMethod includes cat、time-cat、 wilkinson-extended、r-pretty、time、time-pretty、log、pow、quantile、d3-linear}
}
\description{
xAxis settings
}
