% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{yAxis}
\alias{yAxis}
\title{yAxis settins}
\usage{
yAxis(
  gWidget,
  position = NULL,
  label = NULL,
  min = NULL,
  max = NULL,
  minLimit = NULL,
  maxLimit = NULL,
  tickCount = NULL,
  tickInterval = NULL,
  tickMethod = NULL,
  line = NULL,
  tickLine = NULL,
  subTickLine = NULL,
  title = NULL,
  grid = NULL,
  animate = NULL
)
}
\arguments{
\item{animate}{}
}
\value{

}
\description{
yAxis settins
}
