% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes}
\alias{aes}
\title{Title
mapping function just like ggplot's aes().}
\usage{
aes(g, x, y = NULL, color = NULL, size = NULL, shape = NULL, group = NULL)
}
\arguments{
\item{g}{g2plot object}

\item{x}{data mapping for x}

\item{y}{data mapping for y}

\item{color}{the colorField}

\item{size}{the sizeField}

\item{shape}{shapeFiled}

\item{group}{}
}
\value{

}
\description{
Title
mapping function just like ggplot's aes().
}
\examples{
g(data)  \%>\% point()\%>\% aes(x,y,color=z)

}
