% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes}
\alias{aes}
\title{Title
mapping function like ggplot aes()}
\usage{
aes(g, x, y, color = NULL, size = NULL, shape = NULL, group = NULL)
}
\arguments{
\item{x}{data mapping for x}

\item{y}{data mapping for y}

\item{color}{the colorField}

\item{size}{the sizeField}

\item{shape}{shapeFiled}

\item{group}{}
}
\value{

}
\description{
Title
mapping function like ggplot aes()
}
\examples{
g(data) \%>\% aes(x,y,color=z) \%>\% point()

}
