% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{<Add Title>}
\usage{
g(
  data,
  gtype = "scatter",
  mapping = aes(x, y, color, size, shape, type, meta),
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  elementId = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{ylab}{}
}
\description{
<Add Description>
}
