% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{<Add Title>
g2plot function
<Add Description>
g function ,start a plot with g(),also use pipe(%>%) as data transfer.}
\usage{
g(
  data,
  plot_type = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  elementId = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{input data}

\item{plot_type}{}

\item{width}{the width of the plot}

\item{height}{the height of the plot}
}
\description{
<Add Title>
g2plot function
<Add Description>
g function ,start a plot with g(),also use pipe(%>%) as data transfer.
}
