% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{g2label}
\alias{g2label}
\title{visible	boolean	是否显示
type	string	文本标签类型
- line: 跟随在折线尾部
- point: 显示在数据点上
formatter	function	对文本标签内容进行格式化
offsetX	number	在 label 位置的基础上再往 x 方向的偏移量
offsetY	number	在 label 位置的基础上再往 y 方向的偏移量
style	object	配置文本标签样式。}
\usage{
g2label()
}
\description{
visible	boolean	是否显示
type	string	文本标签类型
- line: 跟随在折线尾部
- point: 显示在数据点上
formatter	function	对文本标签内容进行格式化
offsetX	number	在 label 位置的基础上再往 x 方向的偏移量
offsetY	number	在 label 位置的基础上再往 y 方向的偏移量
style	object	配置文本标签样式。
}
