% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend}
\alias{legend}
\title{Title Legend settings for g2plot}
\usage{
legend(
  g,
  legend = TRUE,
  layout = NULL,
  position = NULL,
  title = NULL,
  offsetX = NULL,
  offsetY = NULL,
  background = NULL,
  flipPage = NULL,
  itemWidth = NULL,
  itemHeight = NULL,
  itemName = NULL,
  itemValue = NULL,
  itemSpacing = NULL,
  animate = NULL,
  animateOptio = NULL,
  label = NULL,
  marker = NULL,
  maxWidth = NULL,
  maxHeight = NULL,
  rail = NULL,
  reversed = NULL,
  min = NULL,
  max = NULL,
  value = NULL,
  slidable = NULL,
  track = NULL,
  handler = NULL,
  custom = NULL,
  items = NULL
)
}
\arguments{
\item{g}{object for g2plot}

\item{layout}{layout for legend}

\item{items}{}
}
\value{

}
\description{
Title Legend settings for g2plot
}
\examples{
data<-data.frame(type=c('分类1','分类2','分类3','分类4','分类5','其他'),value=c(1:6))
g(data,plot_type = 'pie') \%>\%pie(radius = 1 )\%>\%g2plot::aes(x='value',y='value',color = 'type') \%>\%legend(position = 'left',layout = 'vertical')
}
