% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(
  data,
  xField = NULL,
  yField = NULL,
  colorField = NULL,
  title = "",
  description = "",
  width = NULL,
  height = NULL,
  forceFit = TRUE,
  padding = NULL,
  pointStyle = NULL,
  color = NULL,
  tooltip = NULL,
  legend = NULL,
  label = NULL,
  quadrant = NULL,
  trendline = NULL,
  theme = NULL,
  events = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{xField}{}

\item{yField}{}

\item{colorField}{}

\item{title}{}

\item{height}{}

\item{forceFit}{}

\item{padding}{}

\item{pointStyle}{}

\item{color}{}

\item{tooltip}{}

\item{legend}{}

\item{label}{}

\item{quadrant}{}

\item{trendline}{}

\item{theme}{}

\item{events}{}
}
\description{
scatter
}
